Add easter egg

+++ js/main.js
@@ -1,6 +1,7 @@
-//Global Counter for "To Do" IDs
+// Global Counter for "To Do" IDs
 let counter = 0;
-
+// Easter Egg Counter
+let easterEgg = 0;
 // Make the template and the Edit section invisible by default
 document.getElementsByClassName("todo-template")[0].style.display = "none";
 // Bottom section can be deleted. Editing now works in the task itself.
@@ -10,29 +11,39 @@ document.getElementsByClassName(
 
 const addItem = () => {
   // Copy the Template
-  let todoTemplate = document.querySelector(".todo-template");
-  let newTask = todoTemplate.cloneNode(true);
-
-  // Assign a new ID to the To Do
-  counter++;
-  let todoID = counter;
-  newTask.id = "todo" + todoID;
-  todoTemplate.after(newTask);
-  newTask.style.display = "flex";
-  console.log(newTask);
-
-  //Parse the Content of the Input into the To Do, then delete the Input
   let todoInput = document.getElementById("create-task__task-input").value;
-  newTask.querySelectorAll(".todo-template__text")[0].textContent = todoInput;
-  document.querySelector("#create-task__task-input").value = "";
+  if (todoInput != "") {
+    let todoTemplate = document.querySelector(".todo-template");
+    let newTask = todoTemplate.cloneNode(true);
 
-  // Add Event Listeners for the Buttons
-  newTask
-    .querySelectorAll("button")[0]
-    .addEventListener("click", editItem.bind(null, todoID));
-  newTask
-    .querySelectorAll("button")[1]
-    .addEventListener("click", deleteItem.bind(null, todoID));
+
+    // Assign a new ID to the To Do
+    counter++;
+    let todoID = counter;
+    newTask.id = "todo" + todoID;
+    todoTemplate.after(newTask);
+    newTask.style.display = "flex";
+    console.log(newTask);
+
+    //Parse the Content of the Input into the To Do, then delete the Input
+    newTask.querySelectorAll(".todo-template__text")[0].textContent = todoInput;
+    document.querySelector("#create-task__task-input").value = "";
+
+    // Add Event Listeners for the Buttons
+    newTask
+        .querySelectorAll("button")[0]
+        .addEventListener("click", editItem.bind(null, todoID));
+    newTask
+        .querySelectorAll("button")[1]
+        .addEventListener("click", deleteItem.bind(null, todoID));
+  } else {
+    easterEgg++;
+    // alert("Please enter a valid string");
+    if (easterEgg === 3) {
+      // Rick Roll
+      window.open("https://www.youtube.com/watch?v=oHg5SJYRHA0");
+    }
+  }
 };
 
 const editItem = (toDoID) => {
